backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# AWS SDK
boto3==1.35.94
botocore==1.35.94

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Data validation
pydantic==2.10.5
pydantic-settings==2.7.1

# WebSocket support
websockets==14.1

# CORS
fastapi-cors==0.0.6

# HTTP client
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Code quality
ruff==0.8.6
mypy==1.14.1
